StringAlgorithms

Documentation for StringAlgorithms.

StringAlgorithms.PalindromicAutomaton

StringAlgorithms.PrefixAutomaton

StringAlgorithms.SuffixAutomaton

StringAlgorithms.longestcommonsubstring

StringAlgorithms.manacher

StringAlgorithms.maximumrotation

StringAlgorithms.minimumrotation

StringAlgorithms.nodeoriented

StringAlgorithms.prefix_function

StringAlgorithms.z_algorithm

StringAlgorithms.PalindromicAutomaton — Method

Create an empty palindromic automaton of key type T.

The Palindromic AutoMaton (PAM, also known as the palindrome tree) is a very useful data structure for solving palindrome-related problems. A detailed introduction to PAM can be found via:

Wikipedia: Palindrome Tree

OI-Wiki: 回文树

To add new values into the automaton, simply use Base.push!.

Getters for the internal fields of the automaton are provided, including:

children(pam): Get the children dictionary of the PAM.

nodecount(pam): Get the number of nodes in the PAM.

lastnodeindex(pam): Get the index of the last node of the PAM.

len(pam, i): Get the palindrome length of the i-th node.

fail(pam, i): Get the fail pointer of the i-th node.

cnt(pam, i): Get the frequency count of the i-th node.

Sample usages

Finding the length of the longest palindromic substring

function longest_palindromic_substring(s)
    pam = PalindromicAutomaton{Char}()
    hi = 0
    for ch in s
        push!(pam, ch)
        hi = max(hi, len(pam, lastnodeindex(pam)))
    end
    return hi
end

longest_palindromic_substring("ddabababacc")

# output

7

StringAlgorithms.PrefixAutomaton — Method

Build an automaton based on the prefix function of a given vector. This is useful when the vector to be searched within is very large.

StringAlgorithms.SuffixAutomaton — Method

Create an empty suffix automaton of key type T.

The Suffix AutoMaton (SAM) is a very useful data structure for solving string-related problems. A detailed introduction to SAM can be found via:

Wikipedia: Suffix Automaton

OI-Wiki: 后缀自动机

StringAlgorithms.longestcommonsubstring — Method

Find the longest common substring/sub-vector of two strings/vectors. When there are more than one with the same length, the first occurring in t will be returned.

StringAlgorithms.manacher — Method

The Manacher algorithm proposed by Glenn K. Manacher in 1975.

Given a vector of length n, so that there will be 2n+1 positions if the intervals between each element are included, returns a vector of length 2n+1 denoting the longest palindrome centered at each position.

Sample usages

Finding the length of the longest palindromic substring

function longest_palindromic_substring(s)
    return maximum(manacher(s))
end

longest_palindromic_substring("ddabababacc")

# output

7

StringAlgorithms.maximumrotation — Method

Find the maximum of all rotations of the given string/vector.

StringAlgorithms.minimumrotation — Method

Find the minimum of all rotations of the given string/vector.

StringAlgorithms.nodeoriented — Method

Transform the internal representation of the suffix automaton into a more user-friendly node-oriented representation, which might be useful when users want to access some internal information.

StringAlgorithms.prefix_function — Method

For each prefix of the given vector, calculate the maximum length such that its nontrivial prefix (not itself) equals to its suffix.

For example, for string ababc,

p[1] = 0

p[2] = 0

p[3] = 1, since the prefix aba has both prefix a and suffix a

p[4] = 2, since the prefix abab has both prefix ab and suffix ab

p[5] = 0

Sample usages

Finding all matches (KMP algorithm)

# The following function finds all occurrences of the pattern in the given vector, including overlapping ones.
function findall(pattern, s)
    n, m = length(s), length(pattern)
    tmp = vcat(collect(pattern), [nothing], collect(s))
    p = prefix_function(tmp)
    return [i-m+1:i for i in 1:n if p[i+m+1] == m]
end

findall("aba", "ddabababacc")

# output

3-element Vector{UnitRange{Int64}}:
 3:5
 5:7
 7:9

StringAlgorithms.z_algorithm — Method

Z Algorithm calculates the longest common prefix of each suffix of the given vector and the whole vector itself. Specially, the Z value of the whole vector is forced to be 0.

Sample usages

Usages of Z Algorithm are almost the same as the prefix function.

Finding all matches

# The following function finds all occurrences of the pattern in the given vector, including overlapping ones.
function findall(pattern, s)
    n, m = length(s), length(pattern)
    tmp = vcat(collect(pattern), [nothing], collect(s))
    z = z_algorithm(tmp)
    return [i:i+m-1 for i in 1:n if z[i+m+1] == m]
end

findall("aba", "ddabababacc")

# output

3-element Vector{UnitRange{Int64}}:
 3:5
 5:7
 7:9